+++ changelog.md
@@ -2,6 +2,8 @@
 
 ### [Latest]
 
+- Add tagger specific cuts to results [!205](https://github.com/umami-hep/puma/pull/205)
+
 ### [v0.2.8] (2023/08/09)
 
 - HL fraction scan plot bugfix [!201](https://github.com/umami-hep/puma/pull/201)